refactor(kernel): fold the sm100_fp8_fp4_gemm_1d1d modules into a subpackage
The three _sm100_fp8_fp4_gemm_1d1d* modules spread the shared 1d1d GEMM
port across the deepgemm/ directory. Collect them into one package:
spec.py (descriptors, heuristics, launch), data.py, and kernel.py.

The package __init__ re-exports spec's public surface, so the five thin
kernel modules keep importing from ._sm100_fp8_fp4_gemm_1d1d unchanged;
only the _data references move to the .data submodule. spec stays a
separate module rather than becoming __init__.py because the license
checker requires every __init__.py to carry the two-line native header,
while spec holds ported DeepGEMM code.

spec.__all__ gains align_up, ceil_div, get_theoretical_mk_alignment, and
gran_k_for, which consumers already imported but the list omitted.

Verified: registry --list --strict, bench_suite --check-imports, and the
deepgemm_sm100_fp8_gemm_1d1d m4096_n576_k7168 config on B200.
