+++ tirx_kernels/deepgemm/k_grouped_fp8_gemm_contiguous.py
@@ -199,7 +199,7 @@ def _spec_for(config: dict):
 
 
 def prepare_data(**config):
-    from ._sm100_fp8_fp4_gemm_1d1d_data import prepare_k_grouped
+    from ._sm100_fp8_fp4_gemm_1d1d.data import prepare_k_grouped
 
     config.pop("label", None)
     return prepare_k_grouped(**config)
@@ -227,7 +227,7 @@ def run_test(**config):
     """Compile, launch and compare against the dequantized-matmul oracle."""
     import torch
 
-    from ._sm100_fp8_fp4_gemm_1d1d_data import assert_within_threshold, calc_diff
+    from ._sm100_fp8_fp4_gemm_1d1d.data import assert_within_threshold, calc_diff
 
     config.pop("label", None)
     data = prepare_data(**config)
@@ -251,7 +251,7 @@ def run_test(**config):
 
 
 def run_bench(*, warmup=None, repeat=None, timer=None, rounds=1, cooldown_s=1.0, **config):
-    from ._sm100_fp8_fp4_gemm_1d1d_data import bench_against_deepgemm, deepgemm_launch_k_grouped
+    from ._sm100_fp8_fp4_gemm_1d1d.data import bench_against_deepgemm, deepgemm_launch_k_grouped
 
     config.pop("label", None)
     data = prepare_data(**config)
