+++ tirx_kernels/deepgemm/m_grouped_fp8_gemm_contiguous.py
@@ -174,7 +174,7 @@ def _spec_for(config: dict):
 
 
 def prepare_data(**config):
-    from ._sm100_fp8_fp4_gemm_1d1d_data import prepare_m_grouped_contiguous
+    from ._sm100_fp8_fp4_gemm_1d1d.data import prepare_m_grouped_contiguous
 
     config.pop("label", None)
     return prepare_m_grouped_contiguous(**config)
@@ -203,7 +203,7 @@ def run_test(**config):
     """Compile, launch and compare against the dequantized-matmul oracle."""
     import torch
 
-    from ._sm100_fp8_fp4_gemm_1d1d_data import assert_within_threshold, calc_diff, psum_slice_diff
+    from ._sm100_fp8_fp4_gemm_1d1d.data import assert_within_threshold, calc_diff, psum_slice_diff
 
     config.pop("label", None)
     data = prepare_data(**config)
@@ -236,7 +236,7 @@ def run_test(**config):
 
 
 def run_bench(*, warmup=None, repeat=None, timer=None, rounds=1, cooldown_s=1.0, **config):
-    from ._sm100_fp8_fp4_gemm_1d1d_data import (
+    from ._sm100_fp8_fp4_gemm_1d1d.data import (
         bench_against_deepgemm,
         deepgemm_launch_m_grouped_contiguous,
     )
